change number from int to float for total bill

+++ tip_calculator.py
@@ -1,6 +1,6 @@
 print("Welcome to the tip calculator! ")
 
-total_bill =int(input ("What was the total bill? "))
+total_bill =float(input ("What was the total bill? "))
 tip_percentage= int(input ("How much tip would you like to give? 10%, 12%, or 15%?"))
 number_of_people= int(input("How many people are splitting the bill? "))
